bug fix
a bug in the models was discovered during the process of revision: the calculation of latent heat that matches the core heat flux was missing a term reflecting the change in liquidus with time. this has been fixed and a better method for calculating the heat flux out of the fluid core has been implemented as well.